requirements.txt
python-telegram-bot>=21.0
langchain-core>=0.2.43
sqlalchemy>=2.0.35
python-dotenv==1.0.1
pydantic>=2.11.5
requests
telegramify-markdown>=1.1.0

llama-index>=0.12.0
llama-index-embeddings-huggingface>=0.4.0
llama-index-vector-stores-qdrant>=0.10.0
qdrant-client>=1.7.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
sentence-transformers>=2.2.0
